+++ package.json
@@ -4,9 +4,12 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "start": "node index.js",
-    "dev": "nodemon index.js"
+    "dev": "nodemon index.js",
+    "coverage": "jest --collectCoverage --",
+    "lint": "eslint",
+    "ci:test": "jest --ci"
   },
   "author": "Pedro Dias",
   "license": "ISC",
@@ -21,6 +24,7 @@
   },
   "devDependencies": {
     "jest": "^27.0.6",
-    "nodemon": "^2.0.12"
+    "nodemon": "^2.0.12",
+    "supertest": "^6.1.5"
   }
 }
