+++ src/controllers/authController.js
@@ -1,14 +1,23 @@
 const router = require('express').Router();
-const authService = require('../services/AuthService');
+const authService = require('../services/authService');
 
 //TODO: middlware used in here should have the req.userId added in the previous middlware
 
 router.get('/cenas', async (req, res, next) => {
-    res.send(await authService.heelo(req, res, next));
+    res.send(await authService.hello(req, res, next));
 })
 
+router.get('/coisas', async (req, res, next) => {
+    res.send(await authService.nice());
+})
+
+
 router.get('/test', async (req, res, next) => {
-    res.send('Test');
+    const data = {
+        user: 'pedro',
+        email: "pedro@gmail.com"
+    }
+    res.send(data);
 })
 
 module.exports = router;
