--- src/services/AuthService.js
@@ -1,7 +1,0 @@
-async function heelo(req, res, next) {
-    return req.body.cenas
-}
-
-module.exports = {
-    heelo
-}
