+++ src/services/authService.js
@@ -1,0 +1,11 @@
+async function hello(req, res, next) {
+    return 'Pedro'
+}
+
+async function nice() {
+    return 'Pedro'
+}
+
+module.exports = {
+    hello, nice
+}
