+++ tests/integration/authController.test.js
@@ -1,0 +1,11 @@
+const request = require('supertest');
+const app = require('../../app');
+
+
+describe('GET /auth', () => {
+    it('responds with json', async () => {
+        const res = await request(app).get('/auth/test');
+
+        expect(res.statusCode).toEqual(200);
+    });
+});
