+++ tests/unit/authService.test.js
@@ -1,0 +1,5 @@
+const authservice = require('../../src/services/authService');
+
+test('return Pedro', async () => {
+    expect(await authservice.nice()).toBe('Pedro');
+});
